feat!: replace int/float with specific-width numeric types and add date/duration (closes #9)
BREAKING CHANGE: The `int` and `float` type keywords are removed.
Use specific-width types instead: i8, u8, i16, u16, i32, u32, i64,
u64, i128, u128, f32, f64.

New types added:
- `date` with ISO 8601 literal syntax: d"2024-03-15"
- `duration` with ISO 8601 literal syntax: dur"P1Y2M3D"

Schema validation now performs range checking per type (e.g. i8
accepts -128..127, u16 accepts 0..65535). Integer values are accepted
by float types via implicit widening.

Runtime adds Value::BigInt(i128) for values exceeding i64 range, and
Value::Date/Value::Duration for the new temporal types. Arithmetic,
comparisons, and all builtins updated for BigInt support. New builtin
functions: date() and duration() with full ISO 8601 validation.

Updated across all bindings (WASM, Python, .NET, JVM, Go, Ruby),
tree-sitter grammar, VS Code/JetBrains/Zed editors, and all
documentation.

+++ bindings/dotnet/tests/Wcl.Tests/WclTests.cs
@@ -326,7 +326,7 @@ namespace Wcl.Tests
         {
             using var doc = WclParser.Parse(@"
                 table users {
-                    name: string  age: int
+                    name: string  age: i64
                     | ""Alice"" | 30 |
                     | ""Bob""   | 25 |
                 }
@@ -506,7 +506,7 @@ server web {
         {
             using var doc = WclParser.Parse(@"
 schema ""Server"" {
-    port: int
+    port: i64
     name: string
 }
 
@@ -515,7 +515,7 @@ server web : Server {
     name = ""web""
 }
             ");
-            // Should have errors — port should be int but got string
+            // Should have errors — port should be i64 but got string
             Assert.True(doc.HasErrors(), "expected schema validation errors");
 
             var errors = doc.Errors();
@@ -527,7 +527,7 @@ server web : Server {
         {
             using var doc = WclParser.Parse(@"
 schema ""Server"" {
-    port: int
+    port: i64
     name: string
 }
